+++ packages/swagger-parser/src/helpers/analyze.ts
@@ -1,4 +1,9 @@
-export const analyze = (spec: any) => {
+import type { SwaggerSpec } from '../types'
+
+/**
+ * This function analyzes a spec and returns a report of the analysis.
+ */
+export const analyze = async (spec: SwaggerSpec) => {
   const hasTitle = spec.info?.title !== undefined
 
   const hasDescription = spec.info?.description !== undefined
